# solutions/pocito/marketplace-server/marketplace-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "marketplace",
    "version": "0.1.0",
    "description": "Requests are scoped by the x-wonder-room header; omitting it uses the marketplace room."
  },
  "paths": {
...
    "/api/v1/tools/": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "List Resources",
        "description": "List every published tool config.",
        "operationId": "list_tools_api_v1_tools__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Tools Api V1 Tools  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Create Resource",
        "description": "Create a tool. 409 on id collision; 422 on invalid body.",
        "operationId": "create_tool_api_v1_tools__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTool"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{id}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Get Resource",
        "description": "Get a single tool config by id. 404 if missing.",
        "operationId": "get_tool_api_v1_tools__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Tool Api V1 Tools  Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tools"
        ],
        "summary": "Update Resource",
        "description": "Partial update. Snapshots the prior state (incl. code bytes), bumps version, audits.",
        "operationId": "update_tool_api_v1_tools__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTool"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Tool Api V1 Tools  Name  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tools"
        ],
        "summary": "Delete Resource",
        "description": "Unpublish + wipe. 204 on success; 404 if the config did not exist.",
        "operationId": "delete_tool_api_v1_tools__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/skills/": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "List Resources",
        "description": "List every published skill manifest.",
        "operationId": "list_skills_api_v1_skills__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Skills Api V1 Skills  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Create Resource",
        "description": "Create a skill. 409 on id collision; 422 on invalid body.",
        "operationId": "create_skill_api_v1_skills__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkill"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/{id}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get Resource",
        "description": "Get a single skill manifest by id. 404 if missing.",
        "operationId": "get_skill_api_v1_skills__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Skill Api V1 Skills  Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "skills"
        ],
        "summary": "Update Resource",
        "description": "Partial update. Snapshots the prior state, bumps version, audits.",
        "operationId": "update_skill_api_v1_skills__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSkill"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Skill Api V1 Skills  Name  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "skills"
        ],
        "summary": "Delete Resource",
        "description": "Unpublish + wipe. 204 on success; 404 if the manifest did not exist.",
        "operationId": "delete_skill_api_v1_skills__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plugins/": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "List Resources",
        "description": "List every published plugin manifest.",
        "operationId": "list_plugins_api_v1_plugins__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Plugins Api V1 Plugins  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "plugins"
        ],
        "summary": "Create Resource",
        "description": "Create a plugin. 409 on id collision; 422 on invalid body.",
        "operationId": "create_plugin_api_v1_plugins__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlugin"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plugins/{id}": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "Get Resource",
        "description": "Get a single plugin manifest. 404 if missing.",
        "operationId": "get_plugin_api_v1_plugins__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Plugin Api V1 Plugins  Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "plugins"
        ],
        "summary": "Update Resource",
        "description": "Partial update. Snapshots the prior state, bumps version, audits.",
        "operationId": "update_plugin_api_v1_plugins__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlugin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Plugin Api V1 Plugins  Name  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "plugins"
        ],
        "summary": "Delete Resource",
        "description": "Unpublish + wipe. 204 on success; 404 if the manifest did not exist.",
        "operationId": "delete_plugin_api_v1_plugins__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/presign/download": {
      "post": {
        "tags": [
          "presign"
        ],
        "summary": "Presign Download",
        "description": "Return a presigned GET URL for payload.key.",
        "operationId": "presign_download_api_v1_presign_download_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Presign Download Api V1 Presign Download Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePlugin": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "display_name",
          "description",
          "config"
        ],
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable semantic identifier used in API paths and references"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Human-readable display name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "hebrew_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hebrew Description"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "title": "Tags"
          },
          "config": {
            "$ref": "#/components/schemas/PluginConfig"
          },
          "readme": {
            "type": "string",
            "default": "",
            "title": "Readme"
          }
        },
        "title": "CreatePlugin",
        "description": "Create body. config.yaml and README.md are uploaded first; the manifest is written last."
      },
...
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail"
          }
        },
        "title": "HTTPValidationError"
      },
...
    }
  }
}